Fix: Making changes to CORS and Allowed Hosts
There are CORS errors appearing on the deployed prod. We have a URL that
is listed in the CORS Origins that is not the actual URL of the site. It
was hard coded but the url of the site is dynamic.

We were also seeing errors in the aws console requesting we add
172.31.35.174 to the allowed hosts, did that.